requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
motor==3.3.2
python-multipart==0.0.9
jinja2==3.1.3
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
aiofiles==23.2.1
Pillow==10.2.0
